setChannels now efficiently fills out an image with a different expression per channel. Used it to simplify and add tests to the color module

+++ src/Geometry.cpp
@@ -595,14 +595,14 @@ void Rotate::parse(vector<string> args) {
 Image Rotate::apply(Image im, float degrees) {
 
     // figure out the rotation matrix
-    double radians = degrees * M_PI / 180;
-    double cosine = cos(radians);
-    double sine = sin(radians);
+    float radians = degrees * M_PI / 180;
+    float cosine = cos(radians);
+    float sine = sin(radians);
     // locate the origin
-    double xorigin = (im.width-1) * 0.5;
-    double yorigin = (im.height-1) * 0.5;
+    float xorigin = (im.width-1) * 0.5;
+    float yorigin = (im.height-1) * 0.5;
 
-    vector<double> matrix(6);
+    vector<float> matrix(6);
     matrix[0] = cosine; matrix[1] = sine; matrix[2] = xorigin - (cosine * xorigin + sine * yorigin);
     matrix[3] = -sine; matrix[4] = cosine; matrix[5] = yorigin - (-sine * xorigin + cosine * yorigin);
     return AffineWarp::apply(im, matrix);
@@ -622,20 +622,20 @@ bool AffineWarp::test() {
 
 void AffineWarp::parse(vector<string> args) {
     assert(args.size() == 6, "-affinewarp takes six arguments\n");
-    vector<double> matrix(6);
+    vector<float> matrix(6);
     for (int i = 0; i < 6; i++) { matrix[i] = readFloat(args[i]); }
     Image im = apply(stack(0), matrix);
     pop();
     push(im);
 }
 
-Image AffineWarp::apply(Image im, vector<double> matrix) {
+Image AffineWarp::apply(Image im, vector<float> matrix) {
 
     assert(matrix.size() == 6, "An affine warp requires a vector with 6 entries\n");
     return apply(im, &matrix[0]);
 }
     
-Image AffineWarp::apply(Image im, double *matrix) {
+Image AffineWarp::apply(Image im, float *matrix) {
     Image out(im.width, im.height, im.frames, im.channels);
 
     vector<float> sample(im.channels);
@@ -1702,7 +1702,7 @@ bool Warp::test() {
     Image a(100, 100, 1, 3);
     Noise::apply(a, 0, 1);
     // Do an affine warp in two ways
-    vector<double> matrix(6);
+    vector<float> matrix(6);
     matrix[0] = 0.9; matrix[1] = 0.1; matrix[2] = 3;
     matrix[3] = -0.2; matrix[4] = 0.8; matrix[5] = 3;
     Image warped = AffineWarp::apply(a, matrix);
